Search-Gen-V/evaluator/eval/pyproject.toml
[project]
name = "nugget-eval"
version = "0.1.0"
description = "Nugget matching evaluation framework"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Slime Team", email = "team@slime.ai"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.60.0",
    "scipy>=1.7.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.scripts]
nugget-eval = "nugget_eval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 119
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 119

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
